+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 - new temporary path for manga images for releases MPL-8 and up
 - new tests for setting up the tree; creating and copying module/bash files and env symlinks
 - added the option for env symlink creation into the setup_tree.py bin file
+- added option to specify default config to write into .version file
 
 Changed
 ^^^^^^^